Install an importable python package with helper functions
There are two motivations for applying this change.

1. It will allow to write unit tests without messing with `PYTHONPATH`
2. It will allow downstream distgit scripts to import these functions
   and reuse them.

I realize that at this point, the added value is negligible (because
no unit tests were added and no dowstream uses the functions) but
this is the first step towards it. At the same time, I deliberately
wanted the change as small as possible and add improvementes once
they are needed, so I tried to not-refactor the code at all even
though I would really love to.